+++ index.html
@@ -228,7 +228,7 @@
   </template>
 
   <script src="js/functions.js"></script>
-  <script src="js/main.js"></script>
+  <script type="module" src="js/main.js" ></script>
 
 </body>
 </html>
